+++ src/main/kotlin/net/ndrei/teslapoweredthingies/common/IFluidStorageMachine.kt
@@ -1,0 +1,4 @@
+package net.ndrei.teslapoweredthingies.common
+
+interface IFluidStorageMachine {
+}