+++ src/main/kotlin/net/ndrei/teslapoweredthingies/machines/portabletank/SimpleTankEntity.kt
@@ -6,13 +6,14 @@ import net.minecraft.tileentity.TileEntity
 import net.minecraftforge.fluids.FluidStack
 import net.minecraftforge.fluids.IFluidTank
 import net.ndrei.teslacorelib.tileentities.SidedTileEntity
+import net.ndrei.teslapoweredthingies.common.IFluidStorageMachine
 import net.ndrei.teslapoweredthingies.render.bakery.SelfRenderingTESR
 
 /**
  * Created by CF on 2017-07-16.
  */
 class SimpleTankEntity
-    : SidedTileEntity(SimpleTankEntity::class.java.name.hashCode()) {
+    : SidedTileEntity(SimpleTankEntity::class.java.name.hashCode()), IFluidStorageMachine {
 
     private lateinit var tank: IFluidTank
 
@@ -22,7 +23,6 @@ class SimpleTankEntity
         this.tank = this.addSimpleFluidTank(24000, "Fluid Tank", EnumDyeColor.BLUE, 20, 24)
     }
 
-    override fun supportsAddons() = false
     override fun canBePaused() = false
     override val allowRedstoneControl: Boolean
         get() = false